+++ DLPrototype/Views/Shared/Navigation.swift
@@ -1,0 +1,24 @@
+//
+//  Navigation.swift
+//  DLPrototype
+//
+//  Created by Ryan Priebe on 2023-07-28.
+//  Copyright © 2023 YegCollective. All rights reserved.
+//
+
+import SwiftUI
+
+public enum Page {
+    case dashboard, today, notes, tasks, projects, jobs, companies
+}
+
+public enum PageGroup: Hashable {
+    case views, entities
+}
+
+public class Navigation: Identifiable, ObservableObject {
+    public var id: UUID = UUID()
+
+    @Published public var view: AnyView?
+    @Published public var parent: Page? = .dashboard
+}
